+ Silent mode.

+++ sound.c
@@ -9,18 +9,23 @@
 
 static char name[NAME_SIZE];
 static Vector processes;
+static int silent;
 
 void initSound(int argc, char *argv[]) {
   if(argc != 1) {
     if(argc == 2) {
-      PlaySound(TEXT(argv[1]), NULL, SND_SYNC | SND_FILENAME);
-      ExitProcess(0);
+      if(strcmp(argv[1], "-s") == 0) {
+        silent = TRUE;
+      } else {
+        PlaySound(TEXT(argv[1]), NULL, SND_SYNC | SND_FILENAME);
+        ExitProcess(0);
+      }
     } else if(argc == 3 && strcmp(argv[1], "LOOP") == 0) {
       PlaySound(TEXT(argv[2]), NULL, SND_LOOP | SND_ASYNC | SND_FILENAME);
+      while(TRUE) Sleep(4294967295);
     } else {
       ExitProcess(0);
     }
-    while(TRUE) Sleep(4294967295);
   } else {
     strcat_s(name, sizeof(name), argv[0]);
     strcat_s(name, sizeof(name), " ");
@@ -28,41 +33,48 @@ void initSound(int argc, char *argv[]) {
 }
 
 Sound PlaySoundNeo(const char path[], int loop) {
-	char args[NAME_SIZE] = "";
-	STARTUPINFOA startInfo = { 0 };
-	PROCESS_INFORMATION processInfo = { 0 };
-	startInfo.cb = sizeof(STARTUPINFOA);
-  strcat_s(args, sizeof(args), name);
-  if(loop) strcat_s(args, sizeof(args), "LOOP ");
-	strcat_s(args, sizeof(args), path);
-	CreateProcessA(NULL, args, NULL, NULL, FALSE, 0, NULL, NULL, &startInfo, &processInfo);
-  CloseHandle(startInfo.hStdInput);
-	CloseHandle(startInfo.hStdOutput);
-	CloseHandle(startInfo.hStdError);
-	CloseHandle(processInfo.hThread);
-  pushAlloc(&processes, sizeof(Sound), &processInfo.hProcess);
-	return processInfo.hProcess;
+  if(!silent) {
+    char args[NAME_SIZE] = "";
+    STARTUPINFOA startInfo = { 0 };
+    PROCESS_INFORMATION processInfo = { 0 };
+    startInfo.cb = sizeof(STARTUPINFOA);
+    strcat_s(args, sizeof(args), name);
+    if(loop) strcat_s(args, sizeof(args), "LOOP ");
+    strcat_s(args, sizeof(args), path);
+    CreateProcessA(NULL, args, NULL, NULL, FALSE, 0, NULL, NULL, &startInfo, &processInfo);
+    CloseHandle(startInfo.hStdInput);
+    CloseHandle(startInfo.hStdOutput);
+    CloseHandle(startInfo.hStdError);
+    CloseHandle(processInfo.hThread);
+    pushAlloc(&processes, sizeof(Sound), &processInfo.hProcess);
+    return processInfo.hProcess;
+  }
+  return NULL;
 }
 
 void StopSound(Sound sound) {
-  Sound *soundPointer;
-  TerminateProcess(sound, 0);
-  CloseHandle(sound);
-  iterf(&processes, &soundPointer) {
-    if(*soundPointer == sound) {
-      removeByData(&processes, soundPointer);
-      free(soundPointer);
-      break;
+  if(!silent) {
+    Sound *soundPointer;
+    TerminateProcess(sound, 0);
+    CloseHandle(sound);
+    iterf(&processes, &soundPointer) {
+      if(*soundPointer == sound) {
+        removeByData(&processes, soundPointer);
+        free(soundPointer);
+        break;
+      }
     }
   }
 }
 
 void deinitSound(void) {
-  Sound *sound;
-  iterf(&processes, &sound) {
-    TerminateProcess(*sound, 0);
-    WaitForSingleObject(*sound, INFINITE);
-    CloseHandle(*sound);
+  if(!silent) {
+    Sound *sound;
+    iterf(&processes, &sound) {
+      TerminateProcess(*sound, 0);
+      WaitForSingleObject(*sound, INFINITE);
+      CloseHandle(*sound);
+    }
+    freeVector(&processes);
   }
-  freeVector(&processes);
 }
